+++ src/Server/Persistence/Entities/EntityProcessor.cs
@@ -39,6 +39,7 @@ public sealed class EntityProcessor
     private const int INDEX_ACCOUNT = 8;
     private const int INDEX_PARENT = 9;
     private const int INDEX_DRAWABLE = 10;
+    private const int INDEX_ANIMATEDSPRITE = 11;
     private readonly IEntityFactory entityFactory;
     private readonly EntityMapper entityMapper;
 
@@ -80,6 +81,7 @@ public sealed class EntityProcessor
         ProcessAccount(reader, builder, entityId);
         ProcessParent(reader, builder);
         ProcessDrawable(reader, builder);
+        ProcessAnimatedSprite(reader, builder);
 
         /* Complete the entity. */
         builder.Build();
@@ -191,6 +193,17 @@ public sealed class EntityProcessor
     {
         if (reader.IsDBNull(INDEX_DRAWABLE)) return;
         if (reader.GetBoolean(INDEX_DRAWABLE)) builder.Drawable();
+    }
+
+    /// <summary>
+    ///     Processes the AnimatedSprite component if present.
+    /// </summary>
+    /// <param name="reader">Reader.</param>
+    /// <param name="builder">Entity builder.</param>
+    private void ProcessAnimatedSprite(IDataReader reader, IEntityBuilder builder)
+    {
+        if (reader.IsDBNull(INDEX_ANIMATEDSPRITE)) return;
+        builder.AnimatedSprite(reader.GetInt32(INDEX_ANIMATEDSPRITE));
     }
 
     /// <summary>